requirements.txt
asgiref==3.7.2
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
async-timeout==4.0.2
PyJWT==1.7.1
redis==4.5.5
